Change packaging to org.apache.trafficcontrol
